requirements.txt
# API Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings>=2.0.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.1

# ML / AI
scikit-learn==1.6.1
xgboost==2.1.3
pandas==2.2.3
numpy>=2.0.0
joblib==1.4.2

# RL Agent
gymnasium==1.0.0
stable-baselines3>=2.6.0

# LLM
google-generativeai==0.8.4

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
python-dateutil==2.9.0.post0

# Authentication
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
email-validator==2.3.0

# Dev
pytest==8.3.4